chore: Refactoring of training tasks
